+++ tst/components/ui/Chat/ChannelList/ChannelItem.test.tsx
@@ -19,6 +19,22 @@ describe('ChannelItem', () => {
     expect(el).toBeInTheDocument();
   });
 
+  it('renders a detailed ChannelItem', () => {
+    const component =
+        <ChannelItem
+            name="Channel 1"
+            displayLastChannelMessageAndTimestamp = {'channel1' === 'channel1'}
+            lastChannelMessage = 'Alice: Hello!'
+            lastChannelMessageTimestamp = "5:01 pm"
+            onClick={() => null} />;
+    const { getByText } = renderWithTheme(lightTheme, component);
+    const el = getByText('Alice: Hello!');
+    expect(el).toBeInTheDocument();
+    const el2 = getByText('5:01 pm');
+    expect(el2).toBeInTheDocument();
+  });
+
+
   it('should render a channel actions button if more options are provided and the channel is active is the list', () => {
     const component = (
       <ChannelList>