+++ CHANGELOG.md
@@ -12,8 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add styling guide documentation for customizing SDK and UI component CSS.
 - Extend and enabled style customizing capabilities on the SDK components.
-- Add a new chat UI component DetailedChannelList which displays a list of channels with channel name, last message content and 
-timestamp, as well as unread channel badge.
+- Updated the ChannelList UI component to take in an optional boolean `displayLastChannelMessageAndTimestamp`. If the boolean is set to true then it will display the last message content and timestamp for each channel item.
 ### Changed
 
 ### Removed